Add Post Entity and In-Memory Datasource
This creates a simple blog post entity with only a title and body fields.  This will hopefully
demonstrate how the loopback framework works in a basic fashion.  This creates a model
(which defines the a model's properties), a controller (which routes incoming requests to
the appropriate method), and a repository (which provides basic CRUD methods to the
in-memory datasource).  

The in-memory datasource is configured to persist all data into a json file in the ./data
directory.  Be sure to create a data dir andgive it the proper permissions.