Add Basic Auth Strategy Provider and Wire to Controller
This creates a basic authentication strategy using passport-http and a new Provider.  Instead of
running the API unprotected and limiting access to it via AWS security groups, we could use this
module to allow the app (and us, remotely) to interact with the API securely.  The authentication
component is one that the Strongloop team is actively developing but is not production-ready as
of the date of this commit.  Either way, it will pay off later to come up with some sort of AAA strategy
now, even if it's just segmenting away the API to only the front end via layer 3.

Find out more here: 
https://github.com/strongloop/loopback-next/blob/master/packages/authentication/README.md