Initial Sandbox Heroku Setup, Build, and Deploy
This switches to a setup where each *.datasource.ts pulls in the relavant secret info from an environement variable, removing the need to use travis encrypt-file and pass decrypted files to deploy steps further down the build pipeline.  When using travis encrypt, be sure to always use --pro (travis-ci.com, not .org) and -r owner/repo so that that CLI encrypts with the correct public key.

Heroku Tokens can be read out using with heroku auth:token and encrypted as well.  Heroku only expects yarn.lock OR package-lock.json, not both.  So, if using yarn, delete before deploying or possibly look at using .slugignore? The deply.run option in the heroku deploy (.travis.yml) is not necessary as npm start will be run by default.  Heroku binds the app to a random port internally, as set by process.env.PORT.  So it is necessary to server.bind('rest.port') on that port at runtime. @loopback/build and @loopback/testlab need to be available to heroku so that lb-tsc and lb-mocha can be run when triggered by npm scripts.  

Lastly, removal of travis encrypt-file and other deploy/build options necessitates removal of old files and decryption/un-tarring steps.