Add Child Model to UserProfile
This adds the User model that implements the UserProfile interface in @loopback/authenticate.
The CURRENT_USER binding is injected into the controller using @inject(user, {optional:true}) for
now to get it working for now.  The verify() method also doesn't actually check the user's credentials
but instead passes along an generic new User.

See this thread for some good ideas:
https://github.com/strongloop/loopback-next/issues/1334